pipecat-starters/gemini_multimodal_live/requirements.txt
pipecatcloud
pipecat-ai[google,daily,silero]
python-dotenv
